pyproject.toml
[tool.poetry]
name = "trader"
version = "0.1.0"
description = "Crypto trading bot."
authors = ["Adam Pisula <adam.pisula@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
apscheduler = "^3.10.4"
asyncio = "^3.4.3"
pydantic = "^2.9.2"
ccxt = "^4.4.14"
python-dotenv = "^1.0.1"
plotly = "^5.24.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
